requirements.txt
Django==5.0.2
seerbit-python-v2==1.0.4
python-dotenv==1.0.1